+++ es6/tests/base.js
@@ -95,7 +95,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 8, 0],
+					currentModuleApiVersion: [3, 9, 0],
 					neededVersion: [5, 6, 0],
 				},
 			}
@@ -110,7 +110,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 8, 0],
+					currentModuleApiVersion: [3, 9, 0],
 					neededVersion: [3, 44, 0],
 				},
 			}